+++ packages/cli/src/editor/components/QuickSwitcher/QuickSwitcherContent.tsx
@@ -9,9 +9,14 @@ import {Internals} from 'remotion';
 import {INPUT_BORDER_COLOR_UNHOVERED} from '../../helpers/colors';
 import {isCompositionStill} from '../../helpers/is-composition-still';
 import {useKeybinding} from '../../helpers/use-keybinding';
+import {
+	makeSearchResults,
+	useMenuStructure,
+} from '../../helpers/use-menu-structure';
 import {ModalsContext} from '../../state/modals';
 import {useSelectComposition} from '../InitialCompositionLoader';
 import {fuzzySearch} from './fuzzy-search';
+import type {Mode} from './NoResults';
 import {QuickSwitcherNoResults} from './NoResults';
 import type {TQuickSwitcherResult} from './QuickSwitcherResult';
 import {QuickSwitcherResult} from './QuickSwitcherResult';
@@ -51,13 +56,38 @@ export const QuickSwitcherContent: React.FC = () => {
 		selectedIndex: 0,
 	});
 	const selectComposition = useSelectComposition();
+	const closeMenu = useCallback(() => undefined, []);
+	const actions = useMenuStructure(closeMenu);
 
 	const {setSelectedModal} = useContext(ModalsContext);
 
 	const keybindings = useKeybinding();
+
+	const mode: Mode = state.query.startsWith('>') ? 'commands' : 'compositions';
+
+	const actualQuery = useMemo(() => {
+		if (mode === 'commands') {
+			return state.query.substring(1).trim();
+		}
+
+		return state.query.trim();
+	}, [mode, state.query]);
+
+	const menuActions = useMemo((): TQuickSwitcherResult[] => {
+		if (mode !== 'commands') {
+			return [];
+		}
+
+		return makeSearchResults(actions, setSelectedModal);
+	}, [actions, mode, setSelectedModal]);
+
 	const resultsArray = useMemo((): TQuickSwitcherResult[] => {
+		if (mode === 'commands') {
+			return fuzzySearch(actualQuery, menuActions);
+		}
+
 		return fuzzySearch(
-			state.query,
+			actualQuery,
 			compositions.map((c) => {
 				return {
 					id: 'composition-' + c.id,
@@ -71,7 +101,14 @@ export const QuickSwitcherContent: React.FC = () => {
 				};
 			})
 		);
-	}, [compositions, state.query, selectComposition, setSelectedModal]);
+	}, [
+		mode,
+		compositions,
+		actualQuery,
+		menuActions,
+		selectComposition,
+		setSelectedModal,
+	]);
 
 	const onArrowDown = useCallback(() => {
 		setState((s) => {
@@ -154,7 +191,7 @@ export const QuickSwitcherContent: React.FC = () => {
 					);
 				})}
 				{resultsArray.length === 0 ? (
-					<QuickSwitcherNoResults query={state.query} />
+					<QuickSwitcherNoResults mode={mode} query={actualQuery} />
 				) : null}
 			</div>
 		</div>
